Code refactor dynamic background component
